# Changed mini-table to use a custom "mode" instead of descriptor's "label."