This is in anticipation of extensions, which will add another flag to "mode."

This slightly improves the performance of serialize.

I do not know why the proto2 numbers changed by 17% in this PR, as this obviously doesn't change proto2 at all. It is unfortunate to know that the signal is this noisy. It's the InitBlock tests specifically that are affected, perhaps the alignment of the initial block plays a significant role.

```
name                                      old time/op  new time/op  delta
ArenaOneAlloc                             22.0ns ± 1%  21.4ns ± 1%   -3.11%  (p=0.000 n=11+12)
ArenaInitialBlockOneAlloc                 6.32ns ± 0%  6.32ns ± 0%     ~     (p=0.449 n=11+12)
LoadDescriptor_Upb                        51.2µs ± 2%  51.6µs ± 1%   +0.87%  (p=0.020 n=12+12)
LoadAdsDescriptor_Upb                     2.63ms ± 0%  2.66ms ± 2%   +1.07%  (p=0.000 n=10+11)
LoadDescriptor_Proto2                      240µs ± 0%   238µs ± 0%   -0.90%  (p=0.000 n=12+12)
LoadAdsDescriptor_Proto2                  12.8ms ± 0%  12.7ms ± 0%   -0.48%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<UseArena,Copy>         10.2µs ± 0%  10.0µs ± 0%   -1.52%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<UseArena,Alias>        8.65µs ± 0%  8.74µs ± 0%   +1.09%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<InitBlock,Copy>        9.70µs ± 0%  9.58µs ± 0%   -1.28%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Alias>       8.23µs ± 0%  8.22µs ± 0%   -0.23%  (p=0.000 n=11+12)
Parse_Proto2<FileDesc,NoArena,Copy>       29.5µs ± 0%  29.5µs ± 0%   +0.11%  (p=0.039 n=12+12)
Parse_Proto2<FileDesc,UseArena,Copy>      20.9µs ± 3%  20.9µs ± 2%     ~     (p=1.000 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     17.0µs ± 1%  19.9µs ± 7%  +16.91%  (p=0.000 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.8µs ± 0%  19.7µs ±10%  +17.06%  (p=0.000 n=12+12)
SerializeDescriptor_Proto2                5.33µs ± 1%  5.33µs ± 0%     ~     (p=0.833 n=12+11)
SerializeDescriptor_Upb                   13.1µs ± 0%  12.3µs ± 0%   -5.68%  (p=0.000 n=12+9)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
   +11%    +320  [ = ]       0    [Unmapped]
  +0.8%     +96  +0.9%     +96    upb/decode.c
    +1.0%     +96  +1.0%     +96    decode_msg
  +0.3%     +48  +0.3%     +48    upb/def.c
    +1.1%     +48  +1.1%     +48    _upb_symtab_addfile
  -0.8%     -48  -0.8%     -48    upb/encode.c
    -1.5%     -48  -1.5%     -48    encode_message
 -10.2%    -416 -12.3%    -416    upb/reflection.c
   -32.9%    -112 -36.8%    -112    upb_msg_get
   -36.1%    -128 -41.0%    -128    upb_msg_whichoneof
   -26.3%    -176 -27.8%    -176    upb_msg_next
  [ = ]       0  -0.2%    -320    TOTAL
```